source/pulse/pyproject.toml
[tool.poetry]
name = "pulse"
version = "0.1.0"
description = "Self-healing mesh network"
authors = ["Jason Blackert <jason.blackert@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
toml = "^0.10.2"
Jinja2 = "^3.1.2"
influxdb = "^5.3.1"
pymodbus = "3.0.0"
pyserial = "^3.5"
numpy = "^1.19.5"
pyroute2 = "^0.7.8"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.2"
pytest = "^7.3.1"
ruff = "^0.0.275"

[tool.poetry.scripts]
pulse = "pulse.pulse:main"

[tool.isort]
profile = "black"

[tool.ruff]
select = ["E", "F", "I", "N", "PTH", "UP"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
